Split code up and integrate properly
